// frontend/src/app/api/models/new-event.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EventType } from '../models/event-type';
export interface NewEvent {

  /**
   * Nur bei Vorgemerkt, ChallengeErhalten und AngebotErhalten erlaubt
   */
  due_on?: string;
  note?: string;
  occurred_on: string;
  type: EventType;
}